remoe to int in cancel call
git-svn-id: https://plugins.svn.wordpress.org/redi-restaurant-reservation/trunk@1025585 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ redi-restaurant-reservation.php
@@ -1020,7 +1020,7 @@ if ( ! class_exists( 'ReDiRestaurantReservation' ) ) {
 
 				case 'cancel':
 					$params = array(
-						'ID'          => (int) self::GetPost( 'ID' ),
+						'ID'          => self::GetPost( 'ID' ),
 						'Email'       => urlencode( self::GetPost( 'Email' ) ),
 						'Reason'      => urlencode( mb_substr( self::GetPost( 'Reason' ), 0, 250 ) ),
 						"Lang"        => str_replace( '_', '-', self::GetPost( 'lang' ) ),
